Add simple grammar for syntax highlighting [#14]
Using a TextMate based grammar for simplicity.

Tree-sitter would be another option, it would however be a lot more
involved.

The current grammar is template-independent. The plan for the future is
to use a dynamically generated grammar based on the active template.

There are currently issues implementing grammar for features which use
nesting (making it non-regular) and for features which use indentation
awareness (making it non-context-free).