Disable buggy typescript-eslint rules [#16]
The rules disabled in this commit (
@typescript-eslint/no-unsafe-assignment,
@typescript-eslint/no-unsafe-call, and
@typescript-eslint/no-unsafe-return), along with previously disabled
@typescript-eslint/no-unsafe-member-access, seem to work incorrectly.
The types of the used constructs here were clearly declared, however
once exported and imported in another module, the eslint plugin somehow
forgets about the types.

Additionally, the before mentioned rule violations go undetected when
running eslint on a built project (see issue #12).

These rules are all from the
@typescript-eslint/recommended-requiring-type-checking config. This
config is not enabled by default when using eslint --init due to it
requiring type checking and thus being slower - therefore it might be
used a lot less frequently than the by default enabled config.

One idea is that these issues might be caused by the usage of
TypeScript's relatively new and not so often used composite projects
feature. This however remains to be explored in more detail in the
future.