Remove specs from dist package [#21]
The tests might rely on dev dependencies which might not be available to
the end user, and to run the specs the user needs to navigate to the dir
of the installed package.

They also increase the package size.

All in all, it seems unnecessary to include them in production.